debug-cleanup-patrolcall
