+++ src/components/Header/Header.js
@@ -7,13 +7,14 @@ import Nav from '../Nav/Nav';
 import ToggleSwitch from '../ToggleSwitch/ToggleSwitch.js';
 
 class Header extends Component {
-
+  
 
   componentDidMount = () => {
     //TO-DO NEED TO CALL THESE DISPATCHES STRATEGICALLY TO ENSURE THEY UPDATE IN REAL TIME
     this.props.dispatch({type: 'FETCH_PATROL'});//get patrol count
     this.props.dispatch({type: 'FETCH_ONCALL'});//get oncall count
     this.props.dispatch({type: 'GET_ACTIVE'}); // dispatch to GET count of all active incidents
+
   }
 
   handleToggle = (event) => {
@@ -29,8 +30,8 @@ class Header extends Component {
       type: 'ADD_CALL_STATUS',
       payload: {onCallValue: !this.props.reduxStore.user.on_call}
     })
-    this.props.dispatch({type: 'FETCH_PATROL'});
-    this.props.dispatch({type: 'FETCH_ONCALL'});
+    // this.props.dispatch({type: 'FETCH_PATROL'});
+    // this.props.dispatch({type: 'FETCH_ONCALL'});
   }
 
   sendPatrolStatus = (param) => {
@@ -38,8 +39,8 @@ class Header extends Component {
       type: 'ADD_PATROL_STATUS',
       payload: {patrolValue: !this.props.reduxStore.user.on_patrol}
     })
-    this.props.dispatch({type: 'FETCH_PATROL'});
-    this.props.dispatch({type: 'FETCH_ONCALL'});
+    // this.props.dispatch({type: 'FETCH_PATROL'});
+    // this.props.dispatch({type: 'FETCH_ONCALL'});
   }
