+++ src/redux/sagas/on.call.saga.js
@@ -1,20 +1,20 @@
-import axios from 'axios';
-import {put, takeEvery } from 'redux-saga/effects';
+// import axios from 'axios';
+// import {put, takeEvery } from 'redux-saga/effects';
 
-function* fetchOnCall() {
-  try{
-    const onCallCount = yield axios.get('/api/oncall');
-    yield console.log('THE PEOPLE ON CALL', onCallCount.data)
-    yield put({type: 'SET_ONCALL', payload: onCallCount.data});
-  }catch(error){
-    console.log('error in fetch oncall', error); 
-  }
-}
-
-
-function* onCallSaga() {
-  yield takeEvery('FETCH_ONCALL', fetchOnCall);
-}
+// function* fetchOnCall() {
+//   try{
+//     const onCallCount = yield axios.get('/api/oncall');
+//     yield console.log('THE PEOPLE ON CALL', onCallCount.data)
+//     yield put({type: 'SET_ONCALL', payload: onCallCount.data});
+//   }catch(error){
+//     console.log('error in fetch oncall', error); 
+//   }
+// }
 
 
-export default onCallSaga;
+// function* onCallSaga() {
+//   yield takeEvery('FETCH_ONCALL', fetchOnCall);
+// }
+
+
+// export default onCallSaga;
